Added playability to the game.
Made it so a random word gets generated and sent to the party leader (first person in the lobby). And he has to draw the word and the others have to guess and get points based on it. After a correct guess the player can no longer guess and also the player that has the word can't draw. (and probably something else added)